requirements.txt
requests
beautifulsoup4
pystray
Pillow 